* Refactoring of Util Code * Code Review remarks
